ModellTog & Multipurpose
ModellTog was a private project developed for a friend who wanted a
program capable of registering his model trains, with possibility to
also register parts (e.g. rails, accessories, figures, etc) and option
to print out to a textfile a wishlist of desired parts with their
registration number.

Multipurpose is a program I developed for myself, with the goal of
having a program able to give me a quick overview of the database
content, create a template file for a ReadMe to attach to a program or
just simply generate incremented text (e.g. for game levels, links,
etc).